tdf#157467 SvxPresetListBox: contextmenu must affect item under mouse
Yikes - especially when delete is involved.
Previously it was deleting the in-use background,
regardless of where the mouse right-click happened.

You might wonder why not just query GetHighlightedItemId?
That is because (at least for GTK)
the highlightedItem is zero'd out as soon as the pop-up starts,
i.e. when the mouse is moved over top of the context menu.

Change-Id: Ib88cb0edb8c2c8f854c35821d8125a1991d12a55
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176447
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>