tdf#162070: enclose delimiters in in formula inside quotes
problem:
1. some characters were not correctly interpreted inside formulas
casuing them to appear as inverted question mark.
2. when some special symbols like 'abs' were inserted as normal string
were replaced with symbols and sometimes also causing to invalidate formula
and add inverted question mark

Note:
So in MSO if you use operator with spaces on either size it will be treated as
literal and when you use them without space in either side it will be operator.
This same thing goes for all operators.

In libre office you will be able to see visual difference
in operator and literal one is slightly bold than other.

Change-Id: Id1b0cda36727c1749619adc858b27212057b37e3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176421
Reviewed-by: Pranam Lashkari <lpranam@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177797
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>