tdf#163340 pasting paragraphs with lists changed again
Pasting paragraph with a list into an empty or completely selected
paragraph with lists has been changed:
If the pasted paragraph(s) are all formatted with a list from the
paragraph style then this list will be preferred.
If some the pasted paragraphs including the first one are formatted
with a hard list then these attributes will be preferred.
If all paragraphs have a hard list or no list at all the list at
the target position will be preferred.

Change-Id: I5aa6834d32f7c69b2eb9f8330b282c0ae6fca559
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176478
Tested-by: Gabor Kelemen <gabor.kelemen.extern@allotropia.de>
Reviewed-by: Thorsten Behrens <thorsten.behrens@allotropia.de>
Tested-by: Jenkins