tdf#161775 vcl: WMF export: remove special case of Dx in EXTTEXTOUT
EMFWriter::ImplWriteTextRecord() has a special case for the last
position of the DX array to write an average width of all the characters
excluding the last one.

This has been the case since CVS initial import, and i was unable to
find any reason for it.

When redacting "Schritte" (large text in bugdoc), it causes the right
edge of the rectangle to be at 2964 instead of 3103, exposing the last
character.

Change-Id: I86391e97c4d06093adf3045b04f69153e8e3b0ef
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176497
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>
Tested-by: Jenkins