tdf#161775 sfx2: SfxRedactionHelper: actually use font for layout
Don't just use the font's height, but set the font at the OutputDevice,
to get the correct Y position of the redaction rectangle.

Still need to increase the top/bottom of the rectangle a little; also add
a little on the right edge, at least the "t" in Arial sticks out.

Change-Id: Ic6505fc7fa8697d72a6b11259005bb308e74a50c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176498
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>