Add minimal vcl weld example application
The "minweld" is a minimal standalone vcl application that uses the
LibreOffice vcl weld framework to display the "tip of the day" dialog on
the screen. Upon clicking on the "Next Tip", a counter is incremented
and the text label containing its value is displayed/updated in the
center of the dialog box.

One can run the application by invoking:

    ./bin/run minweld

This example uses cui/ui/tipofthedaydialog.ui for the UI, but only uses
a few elements of it.

Change-Id: If304c60ddb262182cf23438ed1f83dba85e66391
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176610
Reviewed-by: Hossein   <hossein@libreoffice.org>
Tested-by: Jenkins