tdf#136003 Change labels in the protect sheet dialog
previously a comment `(for unprotected cells)` was added to the `Delete Rows` and `Delete Columns` menu entries as a solution of tdf#136003. The change has been reverted https://gerrit.libreoffice.org/c/core/+/176614

As tdf#136003 says "Can't delete rows, despite of allow to deletes", the source of confusion is these labels `[ ] Delete Rows` and `[ ] Delete Columns` in the protect sheet dialog. Changed these to `Delete rows with unprotected cells` to make it more clear what they are about.

Change-Id: I1009fca2a3cf56ffdc513fdb9b6192b3b965d050
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176616
Tested-by: Jenkins
Reviewed-by: Sahil Gautam <sahil@libreoffice.org>