tdf#163890 sw: fix anchor inserted shape to paragraph
Since the commit 651527b4efe9700c8c8dff58ce5aa86ad5681f16 "sw: fix
double-click opening frame dialog, not graphic dialog on images", when
GetModelPositionForViewPoint is called on an "as-char" graphic it
returns the graphic node. This makes inserted shapes anchor to page as a
fallback since they don't find a text node to anchor to.

With this change, the function returns the anchor text node as before if
the caller explicitly asks for only text frames.

This change breaks testTdf157129 in sw_uiwriter8. After pasting content
ending with a hidden paragraph, SwCursorShell::UpdateCursorPos() should
only move the cursor outside the hidden content. Before this change, the
call to GetLayout()->GetModelPositionForViewPoint didn't change the
cursor position because SwTextCursor::GetModelPositionForViewPoint
called pFlyPor->GetFlyCursorOfst and returned success even when it
failed to get a text node and therefore the cursor didn't change. Now it
finds a valid text node and changes the cursor to that. But this is the
wrong behavior, GetModelPositionForViewPoint should not be called if a
non-hidden paragraph can trivially be found.

Change-Id: I8cf9f0c7dcfe6f4be46435c8506e67606ce8fe9c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176775
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
(cherry picked from commit 7ee34038b12c18b34261beb303f1f945d690761c)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176863
Tested-by: Jenkins