CppunitTest_sw_tiledrendering2: try to make this more stable
This fails from time to time, see
<https://ci.libreoffice.org/job/gerrit_linux_gcc_release/177770/console>
for a case.

When it fails, we get no status callback at all, while the interesting
failure is when we get a status callback, but it's wrong.

Just give up in the rare case when we get no status callback to avoid
false alarms for now.

Change-Id: I8f971b64db878d207f9fd2e4fffc4d4b9a4a14e3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176818
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>