cool#9992 doc electronic sign: make return type for .uno:Signature consistent
It turns out the majority of getCommandValues() return values specify a
commandName and a commandValues key, which is helpful when we don't
track the request for this reply.

Fix .uno:Signature to do the same for consistency.

Change-Id: I46ffe5c36047ef2b59113d0221a1874ba28d335b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177857
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
(cherry picked from commit 0a535147f9bed7856e7ab3513fe3f6c38549a099)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177897