docx: export empty GrabBag elements as self-closing xml tags - 2
In 6cb711645b8d8f3cee0d3454e193cc7bf1cad305 the condition was not
good, because there is a special element called "attributes" which
is not a child element but contains attributes of the current element.
In the condition we should check for 0 elements that are not
"attributes".

Change-Id: I1d6995e13f288902c1cce85ac3b7d4e0da213df8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176907
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
(cherry picked from commit ade846aa29a8dc06a12032702a4dae2158581e89)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176912
Reviewed-by: Andras Timar <andras.timar@collabora.com>
Tested-by: Andras Timar <andras.timar@collabora.com>