docx: export empty GrabBag elements as self-closing xml tags - 2
In 6cb711645b8d8f3cee0d3454e193cc7bf1cad305 the condition was not
good, because there is a special element called "attributes" which
is not a child element but contains attributes of the current element.
In the condition we should check for 0 elements that are not
"attributes".

Change-Id: I1d6995e13f288902c1cce85ac3b7d4e0da213df8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176900
Tested-by: Jenkins
Reviewed-by: Andras Timar <andras.timar@collabora.com>
(cherry picked from commit a1a21ca3d4a8c36620fcb2e2af0b6ddd300f6739)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176918
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>