notebookbar: Pass menu button as parent for its menu
Use the Notebookbar menu button as parent for its
menu. This also makes the button position be used
as the reference.

This allows to drop the now unused
NotebookbarTabControlBase::GetHeaderHeight.

Also, set the reference point to be at the bottom
of the button, so the menu opens below it instead
of approximately in the middle, which matches
how other menu buttons behave.

To test: Set "Tabbed" for the UI variant in
"View" -> "User Interface", then click the
"Hamburger menu" button in the notebookbar.

Change-Id: I0717fad73ff7a42d2bdaaa53a73d055234003d3c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176927
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins