pdf: get rid of PDFWriter::AppendUnicodeTextString
It's not needed when we have COSWriter instead that could be used.
Needed to change IPDFEncryptor from unique_ptr to shared_ptr so we
can share it with COSWriter, but without the need to be set.

Also rename the string literal and unicode methods that take a key
to writeKeyAnd... methods to be mroe explicit. We will need methods
that don't write a key sooner or later.

Change-Id: I15c5d347c6658f7ded2de940acc968dee053d460
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177037
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178763
Tested-by: Tomaž Vajngerl <quikee@gmail.com>
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>