tdf#130857 qt weld: Implement QtInstanceDialog::weld_content_area
When this method gets called the first time, insert
a widget with a QVBoxLayout at the beginning
of the dialog's layout, remember and return that.
On subsequent call, return the same one.

Initially, handle the case where the dialog's
layout is a QBoxLayout (subclass), which is
the case for the "File" -> "Printer Settings"
-> "Options" dialog in Writer.

This should be easy to extend for other layouts
as well when needed. For now, assert when another
layout is used, so it will become clear when
working on adding support for another dialog
that needs this.

Change-Id: Ia41a87f8cf62666efc91c05f25dae5fccb3da41d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177054
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins