tdf#130857 qt weld: Support "Save with password" dialog
Declare support for the "Set Password" dialog that can
be triggered in Writer using "File" -> "Save As", check
"Save with password" checkbox and press "Save" to trigger
the dialog.

This means that native Qt widgets are used for that dialog
now when using the qt5 or qt6 VCL plugin and starting LO with
environment variable SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

This makes use of QtInstanceExpander added in

    Change-Id: I7e3a332c0417b1897ae57d7d4c29609245fb5e19
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Sun Nov 24 01:20:31 2024 +0100

        tdf#130857 qt weld: Add QtInstanceExpander

Currently, I see a small rendering issue when
opening that dialog and expanding the expander:
The "Enter password to allow editing" and
"Confirm password" label text is a bit cropped.
It's shown fine when manually enlarging the dialog a bit,
however.

Change-Id: I37775fa3a39d621696d2a6aaa49bd11d6cfb9350
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177198
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins