tdf#130857 qt weld: Notify about Expander state change
Call weld::Expander::signal_expanded when the
QtExpander expanded state changes, e.g. when
its button gets clicked.

In order to do that, introduce a signal in
QtExpander and connect to that in
QtInstanceExpander.

With this in place, a breakpoint in
weld::Expander::signal_expanded now gets
hit as expected when (un)expanding the
expander in the "Set Password" dialog triggered in Writer
using "File" -> "Save As", check "Save with password"
checkbox and press "Save" when using the qt6 VCL
plugin with SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

This implements what was mentioned as still
missing in earlier commit

    Change-Id: I7e3a332c0417b1897ae57d7d4c29609245fb5e19
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Sun Nov 24 01:20:31 2024 +0100

        tdf#130857 qt weld: Add QtInstanceExpander

as

> Signal handling still needs to be implemented
> (calling `weld::Expander::signal_expanded` when
> the expanded state is toggled).

Change-Id: I9cd1b2cc99018f84ba930d55399953266119bed0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177199
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins