tdf#163486 PVS V581 The conditional expressions of the 'if' statements situated
...alongside each other are identical.

Started when a call inbetween was removed in
e50f6c023926269f76cd73e085b7a246cd0f88fa.

and
V1053 Calling the 'EnableRTL' virtual function in the constructor
...may lead to unexpected result at runtime.

and
V1051 Consider checking for misprints. It's possible that the
...'nTopPos' should be checked here.

V1028 Possible overflow. Consider casting operands of the...
'nStartCol + 1' operator to the 'sal_Int64' type, not the result.

Change-Id: Ided25ab6dc9e9d4a1c72a1b809de586c7b890ce0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177218
Tested-by: Jenkins
Reviewed-by: Aron Budea <aron.budea@collabora.com>
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>