tdf#85976 consider 'contains column labels' property for database range
A database range  is an <table:database-range> element in file  format. It has the
attribute 'table:contains-header'  with values  'true' (default) and  'false', and
the attribute  'table:orientation' with values  'row' (default) and  'column'. The
attributes are only written to file, if the value is not default.

If the  selected range has a  database range element then  use its contains-header
property rather than what was last selected.

Change-Id: I120daa2756c50b6c51fab4f6a0549a0b874abaee
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177231
Reviewed-by: Sahil Gautam <sahil.gautam.extern@allotropia.de>
Tested-by: Jenkins