tdf#124741 export xlsx: no defaultRowHeight without customHeight
Apparently, the value of defaultRowHeight is just some nice
kind of suggestion to Excel, since optimalRowHeight rows
can be less than what is specified in defaultRowHeight.

Since Calc always treats defaultRowHeight as authoritative,
on export it needs to always export it that way.

Apparently there are some TWIPS calculations
going on because while Calc defaults
to exporting a defaultRowHeight of 12.80,
Excel is still round-tripping it as 12.75.

make CppunitTest_sc_subsequent_export_test2 \
    CPPUNIT_TEST_NAME=testTdf120168
(In this unit test, it originates as customHeight=1,
but Calc was round-tripping it without that.
Although the round-trip looked identical in Calc,
in Excel the empty rows were now much shorter,
being height-formatted according to the default font size.)

Change-Id: I7ed6b612de0fa508a89dc4cf357b873de230962c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177276
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>