tdf#147819 doc export: only ShapeFlag::Background if XATTR_FILL
The mere presence of an object somewhere in the document
was causing the document background to force a white color.

This code was developed in the heady days
of figuring out what bits meant what,
so the documentation and code-reading is NOT clear.

make CppunitTest_sw_ww8export \
    CPPUNIT_TEST_NAME=testTdf41542_imagePadding
(My ODT file had obviously spent some of its lifetime as a DOC
because it had a white page background.
I unzipped the file, and hand-adjusted styles.xml to set
fo:background-color="transparent" and draw:fill="none")

Change-Id: I41484ab7f71f4a3156e3193e49127b14e58949cc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177347
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>