Allow toggling grid options on/off via config to update writer
SwGridConfig ConfigItem didn't listen to the config changes.

Note that after this, while toggling e.g. VisibleGrid on/off in expert
settings doesn't initially appear to do a whole pile, it does affect
newly created documents, while before this change it does not.

Existing documents don't update to use that config. Toggling this
on/off via the explicit Show Grid ends up in

SwModule::ApplyUsrPref(const SwViewOption &rUsrPref, SwView* pActView)

with something of a bodge to determine the current SwView in order
to affect the document the setting dialog is launched from, which
is why it changes the current document, and new documents, but not
any other open documents.

Change-Id: Ic57e89e3aa9765ad3c4374ddf02ed718382cf518
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177349
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>