lok: jsdialog: enable font name box after save
When we save - we block UI. In the lok we disable that for
some cases like in commit e005690e9c648ce516c1ed60da6341c7215bf2c9
lok: avoid sending jsdialog messages during background save.

But we also use save file to generate previews of shapes what happens
on eg. resize in Online. Then we trigger disable and enable cycle
for frames - and we end with FontNameBox enabled!

That was causing the JSDialog combobox still being locked - because
frame unlock was using direct VCL API, not weld:: one in the
InterimItemWindow. Let's fix that with proper StateChanged handling.

So now it is functional. But still I think there is some error in the
logic - as after save we should beg the "before save" state of control
not "always enabled". But it's different problem.

Change-Id: Ia8e5c8c864156271b5f68d482234e83eb57b0deb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177444
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>