tdf#130857 gotopagedialog.ui: Resave with glade 3.40
This is the .ui file for the "Edit" -> "Go to page"
dialog in Writer.

This luckily also moves the "page_value" GtkAdjustment
to the top of the file, which (at least for now, in
order to support that dialog) avoids the need to
extend QtBuilder to implement handling of GtkAdjustments
that are referenced in the .ui file before they are
defined.

(If it's still needed in the future, extracting
some of the current logic from VclBuilder into
the BuilderBase or WidgetBuilder base classes for
reuse also by QtBuilder might be an idea.)

Change-Id: Ic03acc1d514d412cd609543754829a370a5678e1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177434
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>