tdf#130857 qt weld: Implement spinbutton cursor/sel methods
Implement QtInstanceSpinButton methods related to
the text cursor and selection.

The QDoubleSpinBox base class QAbstractSpinBox
has a method QAbstractSpinBox::lineEdit [1]
that provides access to the spin box's QLineEdit.
However, that method is protected, and can therefore
not be accessed from QtInstanceBuilder.
Therefore, add corresponding methods to
QtDoubleSpinBox and forward calls to the line edit
from there.

The methods implemented in this commit are
very similar to the corresponding ones in
QtInstanceEntry (which however has has direct access to
the QLineEdit).

[1] https://doc.qt.io/qt-6/qabstractspinbox.html#lineEdit

Change-Id: Ib5d2b55478f88618ed58adc2353824c59015a039
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177436
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>