tdf#130857 qt weld: Also notify about programmatic text changes
Connect to the QLineEdit::textChanged [1] signal instead
of QLineEdit::textEdited [2].

As the doc of the latter says:

> Unlike textChanged(), this signal is not emitted when the text is
> changed programmatically, for example, by calling setText().

Programmatic changes are of interest, however, e.g.
GotoPageDlg::GotoPageDlg connects to the changed signal
of the weld::SpinButton and expects to be notified
when the text is changed by using the spin button
buttons to change the value (and thus the text).

[1] https://doc.qt.io/qt-6/qlineedit.html#textChanged
[2] https://doc.qt.io/qt-6/qlineedit.html#textEdited

Change-Id: Ie19bc852f4ceed0fa79565302975376db7126ea4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177447
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>