tdf#130857 qt weld: Support "Go to Page" dialog
Declare support for the "Edit" -> "Go to Page" dialog
available in e.g. Writer.

This means that native Qt widgets are used for that dialog
now when using the qt5 or qt6 VCL plugin and starting LO with
environment variable SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

One issue seen with the dialog is that by using the
up arrow of the spinbox to increase the value, it
is currently possible to increase the value
beyond the maximum value (last page number),
while this is not the case when typing a number
into the box manually.

This is because the GotoPageDlg::PageModifiedHdl
handler currently only gets called for the
latter case, not the former one, despite

    Change-Id: Ie19bc852f4ceed0fa79565302975376db7126ea4
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Nov 27 22:53:55 2024 +0100

        tdf#130857 qt weld: Also notify about programmatic text changes

and will be addressed in a separate commit.

Change-Id: I1f24cf3925e945ae78a9f1646535e08736cd8786
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177450
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>